lab_2/app/requirements.txt
psycopg2==2.9.5
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3